Description of slots in diffs.
Previously: in the diff windows it was written 'before'/'after' slot: id of slot.
which was unclear.
Now it displays the name of compiler + options.

I had to make a little change to the initialisation process of the web interface,
To allow the object Compiler to see what are the compilers available
as soon as the slots are placed.