Add Conformance viewer
It needs some improvements here and there, plus a good dose of
thinking about the hard-limits actually present, like the max of
6 compilers per view, and 1 view per editor. Maybe we can afford
more compilers per view?
(I think that limiting it to 1 view per editor is a good call tho)

A necessary point before merging would be to fix the fact that
when sharing by URL, the contents of the view are not saved.

Releated issues: #461