Use 'trunk' instead of 'snapshot' to indicate gcc built from source
It's easy to confuse and relate the word 'snapshot' with other things
and it doesn't seem quite communicative. Moreover, we already identify
the compiler built from clang's source as 'trunk'.

So, in order to use clear wordings and to be consistent, use the word
'trunk' (at least in the user interface) to identify the compiler built
from gcc's "trunk" ;-)