Ensure processOptOutput always completes.
Previously if the demangling step thew an exception, we'd never
resolve the promise (or reject it). That meant the compilation request
hung forever awaiting it to finish. We must always ensure new promises
return. "reject" didn't work in this case, so for now I opted to return
the non-demangled output.

Fixes #525

For real this time.