Make existence of 'Short' share link clientside decision
Move checking of google api key into urlshorten-google module so it is
entirely decided in that module if it can shorten or not.

If the urlshorten module returns a falselike object the corresponding
like with dynamically be removed.

This refactor will make it easier to drop in an alternative url
shortening service. It also has the side effect that when google url
shortening fails the 'Short' link will not be shown with a non-shortened
url.