Add versioning for xtensor, xsimd, and xtl
The library xtensor has dependencies on some other QuantStack libraries which are also on CE:
https://github.com/QuantStack/xtensor#dependencies

Currently CE only has the trunk version of all of these, rendering xtensor unusable:
https://godbolt.org/z/0NVe74

Adds versioning for all these libraries so that they can work in harmony.

Aside: not sure if this would be presumptuous, but it might make sense to add [default versions](https://github.com/mattgodbolt/compiler-explorer/blob/master/docs/AddingALibrary.md#setting-default-libraries) for each of them. Of course my work here is selfish as I just want to be able to easily play with xtensor on CE which may not match others' use cases 😄 open to suggestions!