Add rv64gc Clang target
Clang trunk has supported 64-bit RISC-V targets for a while now, but the
default list in compiler-explorer makes it look like that's not the
case.  This patch adds a second RISC-V clang target, targeting rv64gc.

I've also swapped over the default ISA target for the rv32 to rv32gc.
We really expect gc to be the default set of ISA extensions that most
users will be interested in, so that's probably the best bet to show.
I've also explicitly set the ABI here, but that's more of as a reminder
-- the sane ABIs to target are ilp32d and lp64d, but LLVM doesn't
support those yet even in trunk.

I've done absolutely no testing on my end, aside from shoving some
arguemnts into the production site's box and checking to see this works.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>