Use latest GCC toolchain in Clang trunk for C
Latest versions of Clang for C++ and previous versions of Clang for C
point to the latest GCC toolchain at the time of build, so mimic that
behavior for Clang trunk too.

Note that for C it might not be required anyway, since it seems that
commit bc24f262f ("Add clang 7.0 (using latest GCC toolchain for libstdc++")
added it for some reason regarding libstdc++. Still, it is better
to be consistent.