Gracefully handle IO errors in disk cache
There appears to be a race condition between key eviction
and access.  It is possible for the file to be deleted from disk,
but the key to still be in the cache momentarily.

If someone tries to access the key during this period a
file not found error will be thrown, and odd behavior may
occur in the caller.

Instead, we catch/log errors and return a cache miss.