Ensure the tab dropdown shows in front
When there are enough tabs, you use a dropdown selection to pick the
other tabs. This would display under the compiler selection dropdown
list, making it very difficult to use.

The .lm_tab dropdown has its z-index set to 5 to show up on top, but
it's stuck inside the .lm_header stacking context, so it can't render
in front of that. Raising the z-index of that whole stacking context
fixes things.

Fixes #1714