Fix the contrast on some dark-mode icons
This adds some overrides for the hamburger menu styles in dark-mode, as
well as some low-specificity rules for the `fa` icons. They need to be
low specificity in order to avoid overriding colors for icons that are
specifically colored, like the green and red checks and Xs in the status
bars.

Fixes #1720