Fix dark-mode save/load UI
The save/load builds a tabbed list using a .active class to select which
tab to show. This class is also used for styling button highlights in
pickers. We can remove the bare .active and .active:hover rules because
they exactly match the .btn-light.active rules, and every other use of
the active class also has a btn-light class attached to it.

Fixes #1718