Fix empty demangler property detection
In cases where the demangler property is explicitly overriden
to be empty, it will return an empty string, not undefined.

`path.normalize` will then return `.`, resulting in
`Internal Compiler Explorer error: Error: spawn . EACCES` errors.

Having firejail set for executionType somehow masks this issue,
so it was not impacting the live site.