Fix incorrect usage of Promise/async in test suite
In most cases the result was assertions simply not being run.

This was verified by changing the assertion in a way that should
have resulted in failure, and observing no tests failing.

Another case resulted in tests being registered to the wrong suite.

The common theme here is that if you are dealing with async in mocha
you must await everything or return any promise objects you create.

The simplest thing to do is convert the function to use async/await,
except in a `describe` callback where async can *not* be used.  If the
async code is not going to generate any test cases, you can make use of
`before(async () => {});`.  If you are going to dynamically generate any
test cases, you must do so synchronously within the `describe` callback.