Reverts the reversion of the "make" being more prod-like, and fixes it. (#3591)
* Use webpack to autogenerate the files.

That way, we always generate them consistently, be it through
`make` or building for prod. This fixes the `make` issue, and also
means we've further unified the way we build and run with `make`
vs prod.

Unfortunately this doesn't work for `make dev`; which sees the
handlebars content only. That's probably ok though, I think.