Fixcrashissue (#4518)
* await for conan downloads
* await later

This fixes #4512 - essentially any exceptions thrown before the `await` of the build environment would leave a "task" that was trying to write to the temp dir. the temp dir would then be deleted, and the untar would fail, and as nobody was await-ing its return (as the awaits used to be _after_ the throwing code) then we'd get an uncaught exception.