rust: add nightly GCC master build
The Rust frontend for GCC is developed in the github project:
  https://github.com/Rust-GCC/gccrs
This is the compiler denoted by '(GCCRS master)'.

The frontend has been contributed to the main GCC repository and has
been enabled in our nightly builds of the GCC master branch: this is the
compiler denoted by '(GCC master)'.

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>