Reduce the size of tests for bug #1285 (#5561)
As best I can tell the bug happens due to stuff happening _before_ the
user code. So I have removed anything after the user code (ie the giant
gob of dwarf debug info and symbols), in an attempt to reduce the test
sizes in general.

I manually undid the fix from #1285 and the tests correctly fail with
this reduced input.