Sort Armv8 all architectural features clang below trunk version (#5675)
Fixes #5641

The "all architectural features" compiler is trunk, plus a whole bunch
of `-march` flags to enable the kitchen sink for AArch64. Which is
useful for some, but as pointed out in the issue, not representative of
what the average clang on the average system will produce. So plain
trunk is a better first entry for this category.

To fix this, I've changed the semver value so that we will have this
order:
```
(trunk)
(all architectural features, trunk)
17.0.1
<more versions...>
```

`(trunk)` is not treated as a valid semver so it's sorted by the string
value. So the shorter `(trunk)` was shown before the longer `(trunk,
all...`.

This is why the other special compiler versions sort as they do, because
they are `(` then something < `t` (and I didn't find any special
workarounds either).

Putting `trunk` at the end fixes this issue and will not break existing
URLs since the name of the compiler settings file entry (e.g.
`armv8-full-clang-trunk` for C++) has not changed.