rust: update objdumper, linker and defaultCompiler (#5920)
I noticed the default Rust version was outdated, and we didn't have an
upgrade of the `objdumper` and `linker` for a while, so I decided to
send this in advance of tomorrow's Rust version upgrade.

I don't know what is the policy to update `objdumper` and `linker`
though. We have 13.2, 13.1, 12.1, 11.2 and 11.1 around, so I went with
the latest. Hopefully that is OK, please let me know if there is some
kind of rule about it.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>