Add pybind11 library (#5935)
First, thanks for maintaining compiler explorer. It is a wonderful tool.

I have been playing with EDG/reflections on godbolt.org and decided to
make a proof of concept library for python bindings that relies on
reflections. To make my life easier, I wanted to depend on pybind11,
instead of re-learning CPython API.
That brings me to this pull request. I would appreciate having pybind11
available in godbolt.

I am not sure if I am missing anything regarding this. Pybind11 is
header-only and the path specified is the one that should be passed to
`-isystem`.
Also, the issue template mentions tests, but I am not sure how or even
if tests are needed for this change.

I will also understand if you would rather not add more libraries to the
compiler explorer image for the sake of its size.