Fortran: Add LLVM Flang to default compiler config (#6039)
And enable exectution for flang-new.

This assumes that LLVM flang is installed from `ninja install` rather
than a package manager, since I'm not sure any have it yet. So
it's `/usr/local/bin` instead of `/usr/bin`.

gfortran I've assumed is in `/usr/bin` to match the existing
settings in the file.

I removed the `ldPath` settings that the `amazon` config has because:
* Most of them refer to /opt/compiler-explorer.
* Removing the rest didn't prevent me from linking/running binaries.

flang-new used to require a seperate libpgmath library to produce
exectuables but this requirement has been removed so I've enabled
execution.