Update llvm-objdump for Webassembly (#6589)
👋 Hey,

The Webassembly language was recently added (in #6429). I tested it
today on the live enviornment, and it appears to not be working
(example: https://godbolt.org/z/YP5KdscEM).

I don't really know how to debug this, there is no other output other
than `objdump returned 255` which doesn't tell me a whole lot. (If
someone can help me investigate this it would be appreciated)

Looking at the configs for the other languages, they always specify a
objdump binary in the `/opt` path, so maybe `llvm-objdump` is not
available in the system?

This PR attempts to fix it by changing to a fixed version of
`llvm-objdump` in the `/opt` install path. (I've also double checked,
and that version does work locally)